Router: path_slash and path_canon outputs redirects to canonical path
path_* outputs are considered reserved for router's purposes.

When path_slash is set to boolean, terminal slash is added or removed
from path using 302 redirect.

When path_canon is set and does not match the current path, 302 redirect
is issued.

This way the route postprocesor can easily redirect to canonical path.